Leverage Stairway 1.0.0's native context-awareness
MdcHook is greatly reduced and renamed to StairwayLoggingHook: it now only logs at notable state transitions.
